requirements.txt
asn1crypto==0.22.0
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
cryptography>=2.3
django>=1.11.15
djangorestframework==3.9.0
idna==2.6
pycparser==2.18
pyopenssl>=17.5.0
python-dateutil==2.6.1
pytz==2017.2
requests>=2.20.0
six==1.10.0
urllib3>=1.23
django-dbbackup>=3.2.0
django-crontab>=0.7.1
